API Reference

Core functions

smearFEM.assemble_system — Function

assemble_system(ne, NodeList, IEN, ndim, FunctionClass="Q1", nDof=1, ID=nothing, Young=1, ν=0.3)

Assembles the finite element system. # Returns the global stiffness matrix

Arguments:

ne::Interger: number of elements in each direction

NodeList::Matrix{Float64}{ndim,nNodes} : coordinates of the nodes

IEN::Matrix{Int}{nElements,nLocalNodes} : connectivity matrix

ndim::Interger: number of dimensions

nDof::Interger: number of degree of freedom per node

FunctionClass::String: type of basis functions to be considered (Q1:quadratic or Q2:Lagrange)

ID::Matrix{Int}{nNodes,nDof} : matrix that maps the global degrees of freedom to the local degrees of freedom

Young::Float64: Young's modulus

ν::Float64: Poisson's ratio

Returns:

K::SparseMatrixCSC{Float64,Int64}{ndof,ndof} : sparse stiffness matrix

smearFEM.basis_function — Function

basis_function(ξ,η=nothing,ζ=nothing,FunctionClass = "Q1")

Define the basis functions and the gradients for a master element

Arguments:

ξ::Float64: ξ coordinate of the point where the basis function is evaluated

η::Float64: η coordinate of the point where the basis function is evaluated

ζ::Float64: ζ coordinate of the point where the basis function is evaluated

FunctionClass::String: type of basis functions to be considered (Q1:quadratic or Q2:Lagrange)

Returns:

N::Vector{Float64}{,ndof}: basis functions

Delta_N::Matrix{Float64}{ndof,ndim}: gradient of the basis functions

smearFEM.gaussian_quadrature — Function

smearFEM.gaussian_quadrature(a,b,nGaussPoints)

Compute the nodes and weights for the Gaussian quadrature of order 2

Arguments:

a,b::Integer : the limits of the integration interval

nGaussPoints::Integer : number of Gauss points to be considered (2 or 3)

Returns:

ξ::Vector{Float64}{,nGaussPoints}: nodes.

w::Vector{Float64}{,nGaussPoints}: weights

Post processing functions

smearFEM.PlotMesh — Method

PlotMesh(NodeList, IEN)

Function to plot the mesh

Arguments:

NodeList::Matrix{Float64}{nNodes,ndim}: array of nodes.

IEN::Matrix{nElem,nNodes}: IEN array.

smearFEM.back_project — Method

back_project(NodeList, CameraMatrix)

Project the 3D mesh to 2D image plane

Arguments:

NodeList::Matrix{Float64}{3,nbNodes}: 3D mesh grid

CameraMatrix::Matrix{Float64}{3,3}: Camera matrix

Returns:

NodeList2D::Matrix{Float64}{2,nbNodes}: 2D coordinates of the nodes

smearFEM.extract_borders — Function

Extract_borders(NodeList, CameraMatrix, BorderNodesList, state, ne = nothing)

Project the 3D mesh to 2D image plane and extract the border nodes (left and right)

Arguments:

NodeList::Matrix{Float64}{ndim,nNodes} : coordinates of the nodes

CameraMatrix::Matrix{Float64}{3,3} : Camera matrix

BorderNodesList::Vector{Vector{Any}{4,N}: : List of border nodes

state::String : State of the function (init:During the initialization of the mesh or update: when the mesh is updated)

ne::Integer: Number of elements in each direction

Returns:

NodeList::Matrix{Float64}{ndim,nbNodes}: 2D coordinates of the border nodes

smearFEM.fit_curve — Method

fit_curve(border)

Fit a curve to the border nodes of the 2D mesh

smearFEM.inflate_sphere — Method

inflate_sphere(NodeList, x0, x1, y0, y1)

Inflate the sphere to a unit sphere

Arguments:

NodeList::Matrix : {[ndims,nNodes], Matrix{Float64}} : The coordinates of the nodes

x0 : Float64 : The lower bound of the x direction

x1 : Float64 : The upper bound of the x direction

y0 : Float64 : The lower bound of the y direction

y1 : Float64 : The upper bound of the y direction

Returns:

NodeList::Matrix{Float64}{ndims,nNodes}: The coordinates of the nodes after inflation

smearFEM.noramlize — Method

noramlize(q, IEN)

Function normalize the solution vector for plotting

Arguments:

q: solution vector

IEN::Matrix{Float64}{nElem, nNodes}: IEN array

Returns:

qList: normalized list of solutions

smearFEM.truncate_colormap — Function

truncate_colormap(minval=0.0, maxval=1.0, n=100)

Function to truncate a colormap

Arguments:

minval::Integer: minimum value of the colormap

maxval::Integer: maximum value of the colormap

n::Integer: number of colors

Returns:

new_cmap: truncated colormap

smearFEM.write_scene — Method

write_scene(fileName, NodeList, IEN, ne, ndim, fields)

Function to write the solution to a VTK file

Arguments:

fileName::String: name of the VTK file.

NodeList::Matrix{Float64}{nNodes, ndim}: array of nodes.

IEN::Matrix{nElem, nNodes}: IEN array {[nElem, nNodes]}

ne::Integer: number of elements in each direction.

ndim::Integer: number of dimensions

fields::Vector{Vector{Float64}}: solution vector.

smearFEM.write_vtk — Method

write_vtk(fileName, fieldName, NodeList, IEN, ne, ndim, q)

Function to write the solution to a VTK file

Arguments:

fileName::String: name of the VTK file.

NodeList::Matrix[nNodes, ndim]: array of nodes.

IEN::Matrix{Float64}{nElem, nNodes}: IEN array.

ne::Integer: number of elements in each direction.

ndim::Integer: number of dimensions.

q::Vector{Float64}: solution vector.